SVG/requirements.txt
langgraph>=0.3.5         
pydantic>=2.6           
openai>=1.21             
anthropic>=0.22           
jinja2>=3.1               
svgwrite>=1.4             
rapidfuzz>=3.6            
wcag-contrast-ratio>=0.9      
python-pptx>=0.6          
python-dotenv>=1.0        